requirements.txt
python-dotenv==0.21.1
aiogram==2.24
aiohttp==3.8.3
colorlog==6.7.0
APScheduler==3.8.0
asyncpg==0.27.0